move interrupt handlers into kernel
